[SYCL][NFC][SUBGROUPS] Separate half/double type tests from full type tests
