+++ sycl/test/sub_group/broadcast.cpp
@@ -2,10 +2,9 @@
 // CUDA compilation and runtime do not yet support sub-groups.
 
 // RUN: %clangxx -fsycl -fsycl-targets=%sycl_triple %s -o %t.out
-// RUN: %clangxx -fsycl -fsycl-targets=%sycl_triple -D SG_GPU %s -o %t_gpu.out
 // RUN: env SYCL_DEVICE_TYPE=HOST %t.out
 // RUN: %CPU_RUN_PLACEHOLDER %t.out
-// RUN: %GPU_RUN_PLACEHOLDER %t_gpu.out
+// RUN: %GPU_RUN_PLACEHOLDER %t.out
 // RUN: %ACC_RUN_PLACEHOLDER %t.out
 
 //==--------- broadcast.cpp - SYCL sub_group broadcast test ----*- C++ -*---==//
@@ -16,52 +15,8 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "helper.hpp"
-#include <CL/sycl.hpp>
-template <typename T>
-class sycl_subgr;
-using namespace cl::sycl;
-template <typename T>
-void check(queue &Queue) {
-  const int G = 240, L = 60;
-  try {
-    nd_range<1> NdRange(G, L);
-    buffer<T> syclbuf(G);
-    buffer<size_t> sgsizebuf(1);
-    Queue.submit([&](handler &cgh) {
-      auto syclacc = syclbuf.template get_access<access::mode::read_write>(cgh);
-      auto sgsizeacc = sgsizebuf.get_access<access::mode::read_write>(cgh);
-      cgh.parallel_for<sycl_subgr<T>>(NdRange, [=](nd_item<1> NdItem) {
-        intel::sub_group SG = NdItem.get_sub_group();
-        /*Broadcast GID of element with SGLID == SGID */
-        syclacc[NdItem.get_global_id()] =
-            broadcast(SG, T(NdItem.get_global_id(0)), SG.get_group_id());
-        if (NdItem.get_global_id(0) == 0)
-          sgsizeacc[0] = SG.get_max_local_range()[0];
-      });
-    });
-    auto syclacc = syclbuf.template get_access<access::mode::read_write>();
-    auto sgsizeacc = sgsizebuf.get_access<access::mode::read_write>();
-    size_t sg_size = sgsizeacc[0];
-    if (sg_size == 0)
-      sg_size = L;
-    int WGid = -1, SGid = 0;
-    for (int j = 0; j < G; j++) {
-      if (j % L % sg_size == 0) {
-        SGid++;
-      }
-      if (j % L == 0) {
-        WGid++;
-        SGid = 0;
-      }
-      exit_if_not_equal<T>(syclacc[j], L * WGid + SGid + SGid * sg_size,
-                           "broadcasted value");
-    }
-  } catch (exception e) {
-    std::cout << "SYCL exception caught: " << e.what();
-    exit(1);
-  }
-}
+#include "broadcast.hpp"
+
 int main() {
   queue Queue;
   if (!core_sg_supported(Queue.get_device())) {
@@ -73,15 +28,6 @@ int main() {
   check<long>(Queue);
   check<unsigned long>(Queue);
   check<float>(Queue);
-  // broadcast half type is not supported in OCL CPU RT
-#ifdef SG_GPU
-  if (Queue.get_device().has_extension("cl_khr_fp16")) {
-    check<cl::sycl::half>(Queue);
-  }
-#endif
-  if (Queue.get_device().has_extension("cl_khr_fp64")) {
-    check<double>(Queue);
-  }
   std::cout << "Test passed." << std::endl;
   return 0;
 }
