+++ sycl/test/sub_group/reduce_fp64.cpp
@@ -6,16 +6,15 @@
 // RUN: %CPU_RUN_PLACEHOLDER %t.out
 // RUN: %GPU_RUN_PLACEHOLDER %t.out
 // RUN: %ACC_RUN_PLACEHOLDER %t.out
-
-//==--------------- scan.cpp - SYCL sub_group scan test --------*- C++ -*---==//
+///==---------------- reduce_fp64.cpp - SYCL sub_group reduce test ----*- C++ -*---==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
-//===----------------------------------------------------------------------===//
+//===-----------------------------------------------------------------------------===//
 
-#include "scan.hpp"
+#include "reduce.hpp"
 
 int main() {
   queue Queue;
@@ -23,11 +22,7 @@ int main() {
     std::cout << "Skipping test\n";
     return 0;
   }
-  check<int>(Queue);
-  check<unsigned int>(Queue);
-  check<long>(Queue);
-  check<unsigned long>(Queue);
-  check<float>(Queue);
+  check<double>(Queue);
   std::cout << "Test passed." << std::endl;
   return 0;
 }
