+++ sycl/test/sub_group/shuffle_fp16.cpp
@@ -2,32 +2,25 @@
 // CUDA compilation and runtime do not yet support sub-groups.
 //
 // RUN: %clangxx -fsycl -fsycl-targets=%sycl_triple %s -o %t.out
-// RUN: env SYCL_DEVICE_TYPE=HOST %t.out
-// RUN: %CPU_RUN_PLACEHOLDER %t.out
 // RUN: %GPU_RUN_PLACEHOLDER %t.out
-// RUN: %ACC_RUN_PLACEHOLDER %t.out
-
-//==--------------- scan.cpp - SYCL sub_group scan test --------*- C++ -*---==//
+//
+//==------------ shuffle_fp16.cpp - SYCL sub_group shuffle test -----*- C++ -*---==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
-//===----------------------------------------------------------------------===//
+//===--_____--------------------------------------------------------------------===//
 
-#include "scan.hpp"
+#include "shuffle.hpp"
 
 int main() {
   queue Queue;
-  if (!core_sg_supported(Queue.get_device())) {
+  if (!Queue.get_device().has_extension("cl_intel_subgroups")) {
     std::cout << "Skipping test\n";
     return 0;
   }
-  check<int>(Queue);
-  check<unsigned int>(Queue);
-  check<long>(Queue);
-  check<unsigned long>(Queue);
-  check<float>(Queue);
+  check<half>(Queue);
   std::cout << "Test passed." << std::endl;
   return 0;
 }
