+++ sycl/test/sub_group/shuffle_fp64.cpp
@@ -6,28 +6,24 @@
 // RUN: %CPU_RUN_PLACEHOLDER %t.out
 // RUN: %GPU_RUN_PLACEHOLDER %t.out
 // RUN: %ACC_RUN_PLACEHOLDER %t.out
-
-//==--------------- scan.cpp - SYCL sub_group scan test --------*- C++ -*---==//
+//
+//==------------ shuffle_fp64.cpp - SYCL sub_group shuffle test -----*- C++ -*---==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
-//===----------------------------------------------------------------------===//
+//===---------------------------------------------------------------------------===//
 
-#include "scan.hpp"
+#include "shuffle.hpp"
 
 int main() {
   queue Queue;
-  if (!core_sg_supported(Queue.get_device())) {
+  if (!Queue.get_device().has_extension("cl_intel_subgroups")) {
     std::cout << "Skipping test\n";
     return 0;
   }
-  check<int>(Queue);
-  check<unsigned int>(Queue);
-  check<long>(Queue);
-  check<unsigned long>(Queue);
-  check<float>(Queue);
+  check<double>(Queue);
   std::cout << "Test passed." << std::endl;
   return 0;
 }
